Change index to seed.js lint seed.js
