+++ server/index.js
@@ -1,11 +1,14 @@
 const express = require('express');
+const get = require('../db/getPictures.js');
+
 const app = express();
 const port = process.env.PORT || 3000;
 // const db = require('../db/index.js');
 
 // Use express params
 app.get('/:id', (req, res) => {
-  res.send(`Id: ${req.params.id}`);
+  get.getPictures(req.params.id)
+    .then((urls) => res.send(urls));
 });
 
 app.listen(port, () => {
